Timer, Unused IPage
Aded class Timer, which can start and end and be stoped. It aded to Games instead a lot of methods. Was delete interface IPage so classes Games, Lectures and User were rewritten without it. Also method exit was deleted in Games and Lectures, Games instead got rewritten attribute del and new variable game_name, value is name of game as child class.